Refactor class diagram generation tests and extract utility function
- Updated `GenerateClassDiagramCommand.test.ts` to utilize a new helper function `generateClassDiagramForTest` for generating class diagrams, improving code readability and maintainability.
- Moved the diagram generation logic into `testUtils.ts`, which now handles properties, methods, destructured parameters, and default values for testing purposes.
- Ensured all test cases in `GenerateClassDiagramCommand.test.ts` leverage the new utility function for consistent diagram generation.

These changes enhance the structure of the test suite and promote code reuse for class diagram generation logic.
